requirements.txt
bcrypt==3.2.0
boto3==1.16.30
botocore==1.19.30
btcpay-python==1.3.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
ecdsa==0.16.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
geographiclib==1.50
geopy==2.0.0
google-auth==1.23.0
googleapis-common-protos==1.52.0
haversine==2.3.0
idna==2.10
iso3166==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
key-generator==1.0.3
limits==1.5.1
mapbox==0.16.1
mapboxcli==0.8.0
MarkupSafe==1.1.1
msgpack==1.0.0
polyline==1.4.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycoingecko==1.4.0
pycparser==2.20
pycrypto==2.6.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
